app/requirements.txt
asgiref==3.2.10
astroid==2.4.2
Django==3.1.5
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.1.0
psycopg2-binary==2.8.6
pylint==2.5.3
python-dotenv==0.15.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
gunicorn==20.1.0